feat: implementing user and role management

+++ srcs/services/identity/Contracts/Authorization/Roles/IRolesAppService.cs
@@ -5,7 +5,7 @@ using datntdev.Microservice.Srv.Identity.Contracts.Authorization.Roles.Dto;
 namespace datntdev.Microservice.Srv.Identity.Contracts.Authorization.Roles;
 
 public interface IRolesAppService
-    : IAppService<long, RoleDto, RoleCreateDto, RoleUpdateDto>
+    : IAppService<int, RoleDto, RoleCreateDto, RoleUpdateDto>
 {
     Task<PaginatedResult<RoleListDto>> GetAllAsync(PaginatedRequest request);
 }
